// packages/workbook/src/gen/types/workbookCalculationsResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Lemma Workbook API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkbookCalculation } from "./workbookCalculation.js";

export interface WorkbookCalculationsResponse {
  workbookCalculations: WorkbookCalculation[];
  /** @nullable */
  nextCursor: string | null;
}