// packages/workbook/src/gen/types/workbookSnapshot.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Lemma Workbook API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkbookSnapshot {
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  id: string;
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  workbookId: string;
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  calculationId: string;
  /** @minimum 0 */
  snapshotIndex: number;
  createdAt: string;
}